feat: moved intent to common
