+++ apps/ledger-live-mobile/src/mvvm/features/SwapDeviceIntentPOC/intents/broadcastEvmIntent/types.ts
@@ -1,25 +1,17 @@
-import type { Intent, IntentDefinition, IntentPlatformDefinition } from "@ledgerhq/device-intent";
+import type { Intent, IntentPlatformDefinition } from "@ledgerhq/device-intent";
+import type {
+  BroadcastEvmIntentDefinition,
+  BroadcastEvmIntentInput,
+  BroadcastEvmJobState,
+} from "@ledgerhq/live-common/wallet-api/Exchange/intents/broadcastEvm/index";
 
-export type BroadcastEvmJobState =
-  | { type: "broadcasting" }
-  | { type: "broadcasted"; hash: string }
-  | { type: "waiting-receipt"; hash: string; pollCount: number }
-  | { type: "confirmed"; hash: string; blockHeight: number }
-  | { type: "failed"; error: Error };
-
-export type BroadcastEvmIntentInput = {
-  /** Fully signed (RLP-serialised) EVM transaction hex, as produced by the sign intent. */
-  signedTxHex: string;
-  /** Currency id of the parent EVM chain (e.g. `"ethereum"`). */
-  currencyId: string;
+export type {
+  BroadcastEvmIntentDefinition,
+  BroadcastEvmIntentInput,
+  BroadcastEvmJobState,
 };
 
 export type BroadcastEvmIntentExtraProps = Record<string, never>;
-
-export type BroadcastEvmIntentDefinition = IntentDefinition<
-  BroadcastEvmJobState,
-  BroadcastEvmIntentInput
->;
 
 export type BroadcastEvmIntentPlatformDefinition = IntentPlatformDefinition<
   BroadcastEvmJobState,
