+++ apps/ledger-live-mobile/src/mvvm/features/SwapDeviceIntentPOC/intents/signSwapEvmIntent/intentLWMDefinition.ts
@@ -1,5 +1,5 @@
+import { signSwapEvmIntentDefinition } from "@ledgerhq/live-common/wallet-api/Exchange/intents/signSwapEvm/index";
 import { SignSwapEvmIntentComponentLWM } from "./componentLWM";
-import { signSwapEvmIntentDefinition } from "./intentDefinition";
 import type { SignSwapEvmIntentPlatformDefinition } from "./types";
 
 export const signSwapEvmIntentLWMDefinition: SignSwapEvmIntentPlatformDefinition = {
