+++ libs/ledger-live-common/src/wallet-api/Exchange/intents/broadcastEvm/index.ts
@@ -1,0 +1,7 @@
+export { broadcastEvmIntentDefinition } from "./intentDefinition";
+export { broadcastEvmJob } from "./job";
+export type {
+  BroadcastEvmIntentDefinition,
+  BroadcastEvmIntentInput,
+  BroadcastEvmJobState,
+} from "./types";
