+++ libs/ledger-live-common/src/wallet-api/Exchange/intents/index.ts
@@ -1,0 +1,31 @@
+/**
+ * Cross-platform device intents used by the wallet-side `custom.swap` flow.
+ *
+ * Each intent has a `Job` that runs against a connected device (or the
+ * network, in the broadcast case) and emits typed `JobState` values. The
+ * platform-specific React components live alongside their callers (LWM
+ * under `apps/ledger-live-mobile`, LWD under `apps/ledger-live-desktop`).
+ */
+export {
+  signApprovalEvmIntentDefinition,
+  signApprovalEvmJob,
+} from "./signApprovalEvm";
+export type {
+  SignApprovalEvmIntentDefinition,
+  SignApprovalEvmIntentInput,
+  SignApprovalEvmJobState,
+} from "./signApprovalEvm";
+
+export { signSwapEvmIntentDefinition, signSwapEvmJob } from "./signSwapEvm";
+export type {
+  SignSwapEvmIntentDefinition,
+  SignSwapEvmIntentInput,
+  SignSwapEvmJobState,
+} from "./signSwapEvm";
+
+export { broadcastEvmIntentDefinition, broadcastEvmJob } from "./broadcastEvm";
+export type {
+  BroadcastEvmIntentDefinition,
+  BroadcastEvmIntentInput,
+  BroadcastEvmJobState,
+} from "./broadcastEvm";
