+++ libs/ledger-live-common/src/wallet-api/Exchange/intents/signApprovalEvm/index.ts
@@ -1,0 +1,7 @@
+export { signApprovalEvmIntentDefinition } from "./intentDefinition";
+export { signApprovalEvmJob } from "./job";
+export type {
+  SignApprovalEvmIntentDefinition,
+  SignApprovalEvmIntentInput,
+  SignApprovalEvmJobState,
+} from "./types";
