+++ libs/ledger-live-common/src/wallet-api/Exchange/intents/signApprovalEvm/intentDefinition.ts
@@ -3,11 +3,10 @@ import type { SignApprovalEvmIntentDefinition } from "./types";
 
 /**
  * Cross-platform definition for the EVM token-approval signing intent used by
- * the swap POC.
+ * the swap flow.
  *
- * Mirrors the demo DMK signer intent: the device is required and lock-state
- * handling is left to the job so the device-action's own UI states can be
- * surfaced through {@link SignApprovalEvmJobState}.
+ * The device is required and lock-state handling is left to the job so the
+ * device-action's own UI states can be surfaced through {@link SignApprovalEvmJobState}.
  */
 export const signApprovalEvmIntentDefinition: SignApprovalEvmIntentDefinition = {
   label: "Sign EVM token approval",
