+++ libs/ledger-live-common/src/wallet-api/Exchange/intents/signSwapEvm/index.ts
@@ -1,0 +1,7 @@
+export { signSwapEvmIntentDefinition } from "./intentDefinition";
+export { signSwapEvmJob } from "./job";
+export type {
+  SignSwapEvmIntentDefinition,
+  SignSwapEvmIntentInput,
+  SignSwapEvmJobState,
+} from "./types";
