+++ libs/ledger-live-common/src/wallet-api/Exchange/intents/signSwapEvm/intentDefinition.ts
@@ -2,8 +2,8 @@ import { signSwapEvmJob } from "./job";
 import type { SignSwapEvmIntentDefinition } from "./types";
 
 /**
- * Cross-platform definition for the EVM swap signing intent used by the
- * swap POC. Mirrors {@link signApprovalEvmIntentDefinition}.
+ * Cross-platform definition for the EVM swap signing intent. Mirrors
+ * {@link signApprovalEvmIntentDefinition}.
  */
 export const signSwapEvmIntentDefinition: SignSwapEvmIntentDefinition = {
   label: "Sign EVM swap",
